Move hitbox from utils to drawing
Move hitbox from utils to drawing
