+++ src/drawing/hitbox/index.test.ts
